Safari-609.1.11.1/LayoutTests/media/video-inactive-playback-expected.txt


RUN(internals.setMediaSessionRestrictions("videoaudio", ""))
RUN(video.src = findMediaFile("video", "content/test"))
EVENT(canplaythrough)
RUN(video.play())
EVENT(playing)
RUN(internals.applicationWillBecomeInactive())
EXPECTED (video.paused == 'false') OK
RUN(internals.setMediaSessionRestrictions("videoaudio", "inactiveprocessplaybackrestricted"))
RUN(internals.applicationWillBecomeInactive())
EVENT(pause)
RUN(internals.applicationDidBecomeActive())
EVENT(playing)
END OF TEST
